Move TUI config into its own file
